[Refactor] PaymentService 리팩토링 및 주문/상품/배송 스키마 개선
- PaymentService: BillingService 명칭 변경 및 PaymentMaster 엔티티 적용

- OrderDetail: 배송/클레임 필드 추가 및 Audit 컬럼 적용

- Product: BrandName 필드 추가 및 Audit 컬럼 보완

- Delivery: 주문번호/상세ID/택배사명 필드 추가 및 Audit 컬럼 적용 (CourierCompany 삭제)

- Database: init_db.sql 및 개별 SQL 파일 동기화

+++ src/main/java/com/nanum/admin/manager/repository/ManagerMenuRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ManagerMenuRepository extends JpaRepository<ManagerMenu, Integer> {
+public interface ManagerMenuRepository extends JpaRepository<ManagerMenu, Long> {
 }
